requirements.txt
aws-cdk-lib==2.133.0
constructs>=10.0.0,<11.0.0
pyyaml==6.0.1
pytest==6.2.5
cdk-nag